pyproject.toml
[project]
name = "yamdb-api"
version = "1.0.0"
description = "Веб ресурс для сбора отзывов пользователей на произведения"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.2",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-nested-routers>=0.95.0",
    "drf-spectacular>=0.29.0",
    "python-dotenv>=1.2.1",
]

[tool.ruff]
line-length = 100
indent-width = 4
target-version = "py314"

# Exclude a variety of commonly ignored directories.
exclude = [
  ".bzr", ".direnv", ".eggs", ".git", ".git-rewrite", ".hg", ".ipynb_checkpoints", ".mypy_cache", ".nox",
  ".pants.d", ".pyenv", ".pytest_cache", ".pytype", ".ruff_cache", ".svn", ".tox", ".venv", ".vscode",
  "__pypackages__", "_build", "buck-out", "build", "dist", "node_modules", "site-packages", "venv", "migrations"
]

[tool.ruff.lint]
select = [
  "E", "W",  # * pycodestyle
  "F",  # * Pyflakes
  "PL",  # * Pylint
  "UP",  # * pyupgrade
  "B",  # * flake8-bugbear
  "SIM",  # * flake8-simplify
  "I",  # * isort
  "FAST",  # * FastAPI
  "NPY",  # * NumPy-specific rules
  "PD",  # * pandas-vet
  "N",  # * pep8-naming
  "DOC",  # * pydoclint
]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"
